+++ packages/twopoint5d/src/stage/IGetRenderPass.ts
@@ -1,8 +1,0 @@
-import type {RenderPass} from 'three/addons/postprocessing/RenderPass.js';
-
-export interface IGetRenderPass {
-  getRenderPass(): RenderPass;
-}
-
-export const hasGetRenderPass = (obj: object): obj is IGetRenderPass =>
-  obj && typeof (obj as IGetRenderPass).getRenderPass === 'function';