+++ packages/twopoint5d/CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- refactor `IStage`, `PostProcessingRenderer`, add `Stage2DRenderPass`
+
 &mldr;
+
 
 ## [0.11.0] - 2025-04-26